+++ docs/conf.py
@@ -243,7 +243,7 @@ texinfo_documents = [
 
 
 # Example configuration for intersphinx: refer to the Python standard library.
-intersphinx_mapping = {'http://docs.python.org/': None}
+intersphinx_mapping = {'https://docs.python.org/': None}
 
 # List of items to ignore when running linkcheck
 linkcheck_ignore = [