+++ CHANGES.rst
@@ -4,9 +4,11 @@
 0.4 (unreleased)
 ----------------
 
-- Add support for Python 3.9.
+- Add support for Python 3.9, and 3.10.
 
 - Remove DeprecationWarning from using ``collections.abc``.
+
+- Move tests to GitHub actions.
 
 0.3 (2020-08-10)
 ----------------
@@ -15,7 +17,7 @@
 
 - Drop support for Python 2.6, 3.2, 3,3, and 3.4.
 
-- Add support for Python 3.4 and PyPy3.
+- Add support for PyPy3.
 
 0.2 (2013-10-21)
 ----------------